# openapi.json
{"openapi":"3.1.0","info":{"title":"Audio procesing app","summary":"This service allows for the upload, download and management of audio files","description":"\nThis service contains several endpoints to manage audio files that are described in this documentation.\nTo use all the endpoins within the '/audio' path an api key is required. Please be sure that your requests contain a valid api key.\n","version":"0.1.0"},"paths":{"/audio/upload":{"post":{"summary":"Audio Upload","operationId":"audio_upload_audio_upload_post","parameters":[{"name":"x-api-key","in":"header","required":true,"schema":{"type":"string","title":"X-Api-Key"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_audio_upload_audio_upload_post"}}}},"responses":{"201":{"description":"Successful Response"},"403":{"Forbidden":401,"description":"Forbidden"},"415":{"description":"Unsupported Media Type"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/audio/{file_id}/download":{"get":{"summary":"Audio Download","operationId":"audio_download_audio__file_id__download_get","parameters":[{"name":"file_id","in":"path","required":true,"schema":{"type":"string","title":"File Id"}},{"name":"x-api-key","in":"header","required":true,"schema":{"type":"string","title":"X-Api-Key"}}],"responses":{"200":{"description":"Success.","content":{"audio/*":{}}},"403":{"Forbidden":401,"description":"Forbidden"},"404":{"description":"Not Found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/audio/{file_id}/adjust_volume":{"put":{"summary":"Adjust Volume","description":"Please provide a value for the volume increase in decibels.     A positive value will increase the volumen whereas a negative one will decrease it.     This endpoint only updates the volume and does not download the file, to access the updated file, please use the download endpoint","operationId":"adjust_volume_audio__file_id__adjust_volume_put","parameters":[{"name":"file_id","in":"path","required":true,"schema":{"type":"string","title":"File Id"}},{"name":"x-api-key","in":"header","required":true,"schema":{"type":"string","title":"X-Api-Key"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VolumeAdjustRequest"}}}},"responses":{"204":{"description":"Successful Response"},"403":{"Forbidden":401,"description":"Forbidden"},"404":{"description":"Not Found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/healthz":{"get":{"summary":"Health Check","operationId":"health_check_healthz_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Body_audio_upload_audio_upload_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_audio_upload_audio_upload_post"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"VolumeAdjustRequest":{"properties":{"volume_increase":{"type":"integer","title":"Volume Increase"}},"type":"object","required":["volume_increase"],"title":"VolumeAdjustRequest"}}}}